@@ -1,24 +1,19 @@
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.FeatureManagement;
-using System.Threading.Tasks;
+namespace AzureAppConfigWebApp.Pages;
 
-namespace AzureAppConfigWebApp.Pages
+public class FeatureSampleModel : PageModel
 {
-    public class FeatureSampleModel : PageModel
+    private readonly IFeatureManager _featureManager;
+    public FeatureSampleModel(IFeatureManager featureManager)
     {
-        private readonly IFeatureManager _featureManager;
-        public FeatureSampleModel(IFeatureManager featureManager)
-        {
-            _featureManager = featureManager;
-        }
+        _featureManager = featureManager;
+    }
 
-        public string? FeatureXText { get; private set; }
+    public string? FeatureXText { get; private set; }
 
-        public async Task OnGetAsync()
-        {
-            bool featureX = await _featureManager.IsEnabledAsync("FeatureX");
-            string featureText = featureX ? "is" : "is not";
-            FeatureXText = $"FeatureX {featureText} available"; 
-        }
+    public async Task OnGetAsync()
+    {
+        bool featureX = await _featureManager.IsEnabledAsync("FeatureX");
+        string featureText = featureX ? "is" : "is not";
+        FeatureXText = $"FeatureX {featureText} available";
     }
 }
