@@ -1,8 +1,4 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using System;
-using Microsoft.Extensions.Hosting;
-
-using var host = Host.CreateDefaultBuilder()
+﻿using var host = Host.CreateDefaultBuilder()
     .ConfigureServices(services =>
     {
         // services.AddOptions(); // already added from host
@@ -17,4 +13,3 @@ using var host = Host.CreateDefaultBuilder()
 var controller = host.Services.GetRequiredService<HomeController>();
 string result = controller.Hello("Katharina");
 Console.WriteLine(result);
-
